+++ client/src/components/MainPage/map/map.module.css
@@ -3,8 +3,8 @@
 .category {position:absolute;overflow:hidden;top:10px;right:50px;width:220px;height:60px;z-index:10;border:1px solid black; border-radius: 5px; font-family:'Malgun Gothic','맑은 고딕',sans-serif;font-size:11px;text-align:center;background-color:#fff;}
 .category .menu_selected {background:rgba(0, 0, 0, 0.5);color:#fff;border-left:1px solid #915B2F;border-right:1px solid #915B2F;margin:0 -1px;} 
 .category li{list-style:none;float:left;width:55px;height:60px;padding-top:5px;cursor:pointer;} 
-.category .ico_comm {display:block;margin:0 auto 2px;width:50px;height:40px;background:url('../../imgs/markerIconSprites.png') no-repeat;} 
-.category .ico_all {background:url('../../imgs/all_icon.png') no-repeat; background-position:5px 0;}
+.category .ico_comm {display:block;margin:0 auto 2px;width:50px;height:40px;background:url('../../../imgs/markerIconSprites.png') no-repeat;} 
+.category .ico_all {background:url('../../../imgs/all_icon.png') no-repeat; background-position:5px 0;}
 .category .ico_cctv {background-position:-127px -10px;}  
 .category .ico_shelter {background-position:-66px -10px;}   
 .category .ico_report {background-position:-5px -10px;}
